% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbci_targets_template.R
\name{add_project}
\alias{add_project}
\title{Add a new targets 'project'}
\usage{
add_project(project)
}
\arguments{
\item{project}{Character. The name of the project. Do not include .R.}
}
\description{
Add a new targets 'project'
}
\details{
This function will produce the following files:

./R/targets/project.R
./R/functions/project.R

It will also update () the make.R script to include this new project, and add
it to the \verb{_targets.yaml} file via tar_config_set. Note that if not already
created, running this script will create another directory, \code{stores}, which
will contain another subdirectory of the name \code{project}.
}
