% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_rna_conc}
\alias{dummy_rna_conc}
\title{Example data of RNA samples with concentrations}
\format{
A data frame with 8 rows and 2 columns
\describe{
\item{sample}{name of sample}
\item{conc}{concentration of RNA, in ng/uL}
}
}
\usage{
dummy_rna_conc
}
\description{
A dataset containing fabricated sample names and RNA concentrations
}
\keyword{datasets}
