% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_heatmap.R
\name{eda_heatmap}
\alias{eda_heatmap}
\title{Make a heatmap of a select number of genes}
\usage{
eda_heatmap(dds, genes, stratifier = NULL, assay = 2)
}
\arguments{
\item{dds}{a \code{dds} object with HGCN symbols as rownames}

\item{genes}{a character vector containing the names of HGCN symbols to be plotted in the heatmap}

\item{stratifier}{optional. A character vector specifying a column in colData by which to color samples in the pheatmap.}

\item{assay}{the assay number that contains normalized counts}
}
\value{
a \code{pheatmap}
}
\description{
Make a heatmap of a select number of genes
}
