% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_one_gene.R
\name{eda_one_gene}
\alias{eda_one_gene}
\title{Plot the expression of one gene across a stratifying argument}
\usage{
eda_one_gene(dds, gene, stratifier, assay = 2)
}
\arguments{
\item{dds}{A \code{DESeqDataSet} object}

\item{gene}{A character vector HGNC Symbol, like "KRT14"}

\item{stratifier}{A name of a column of the \code{dds} \code{colData} that will be used as the x-axis to stratify gene expression into groups}

\item{assay}{The index of the assay that contains the normalized counts. Defaults to \code{2}.}
}
\value{
A ggplot
}
\description{
Plot the expression of one gene across a stratifying argument
}
