% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evos.R
\name{evos_flatten}
\alias{evos_flatten}
\title{Flatten an EVOS timelapse directory}
\usage{
evos_flatten(path, parse_names = FALSE)
}
\arguments{
\item{path}{character. Path to directory that contains all timelapses}
}
\value{
Invisibly returns logical vector the length of the number of files.
As a side effect, all files in that directory will be copied to the top
level of that directory with a file name that is a concatination of all
parent directories, separated by an underscore.
}
\description{
Flatten an EVOS timelapse directory
}
