% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evos.R
\name{evos_parse_name}
\alias{evos_parse_name}
\title{Sensibly parse default EVOS file names}
\usage{
evos_parse_name(name)
}
\arguments{
\item{names}{}
}
\value{
A filename that takes the form of
"parent-directory/b{N}_{YYYY-MM-DD}<_filter>.tif" Where {N} is the number
of the beacon and <_filter> is an underscore prepended filter name (eg
_phase) which may or may not exist.
}
\description{
Sensibly parse default EVOS file names
}
