% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{fit_mtt}
\alias{fit_mtt}
\title{Model curve, produce plots, and calculate IC per MTT condition}
\usage{
fit_mtt(df, drug_conc, ic_pct)
}
\arguments{
\item{df}{a \code{data.frame} containing a \code{condition}, \code{div}, and \code{drug} column}

\item{drug_conc}{a numeric vector containing drug concentrations of the
conditions, from left to right}

\item{ic_pct}{numeric. The \%IC desired, where 25 would represent the
concentration at which growth was reduced by 25\% vs baseline}
}
\value{
A \code{tibble}
}
\description{
Model curve, produce plots, and calculate IC per MTT condition
}
