% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{get_best_factor}
\alias{get_best_factor}
\title{Calculate a sensible dilution factor}
\usage{
get_best_factor(vol_to_add)
}
\arguments{
\item{vol_to_add}{numeric. 'Naive' volume to add, before dilution.}
}
\value{
an integer, either 1 (no dilution) or something divisible by 5.
}
\description{
If the volume of RNA to add is < 1uL, it must be diluted. Dilutions are easy
to calculate in one's head only if they are integers (divisible by 5
preferred). Further, this dilution should be as small as reasonably possible,
otherwise it will become too dilute.
}
\keyword{internal}
