% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gbci}
\alias{get_gbci}
\alias{get_gbci_file}
\alias{get_gbci_dir}
\title{Get a directory or file from the GBCI SharePoint}
\usage{
get_gbci(path, dest = NULL, overwrite = FALSE, create_dir = TRUE, drive = NULL)

get_gbci_file(
  path,
  dest = NULL,
  overwrite = FALSE,
  create_dir = TRUE,
  drive = NULL
)

get_gbci_dir(
  path,
  dest = NULL,
  overwrite = FALSE,
  create_dir = TRUE,
  drive = NULL
)
}
\arguments{
\item{path}{Path to file or directory on SharePoint}

\item{dest}{Where to put the file, and what to name it. Defaults to
tempfile/dir}

\item{overwrite}{Logical. Should files be overwritten if they aready exist?}

\item{create_dir}{Logical. Should the destination directory be created if it
doesn't exist?}

\item{drive}{Optional. A \code{Microsoft365R::ms_drive}. Can be passed from parent
functions to avoid multiple calls, which can be faster.}
}
\value{
The destination of the file or directory (\code{dest})
}
\description{
This function will call either \code{get_gbci_file} or \code{get_gbci_dir} downstream
depending on the path argument.
}
\examples{
\dontrun{
get_gbci_dir("Raw Data/SPECTRAmax/aragaki-kai/", "path/to/my/dir")
}
}
