% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gbci_dir}
\alias{get_gbci_dir}
\title{Recursively download a directory from SharePoint}
\usage{
get_gbci_dir(path, dest)
}
\arguments{
\item{path}{Path to the directory on SharePoint}

\item{dest}{Path to where the file should be downloaded}
}
\value{
Returns \code{dest}
}
\description{
Recursively download a directory from SharePoint
}
\examples{
\dontrun{
get_gbci_dir("Raw Data/SPECTRAmax/aragaki-kai/", "path/to/my/dir")
}
}
