% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{get_ic}
\alias{get_ic}
\title{Get the IC value of a fit at a given percent}
\usage{
get_ic(fit, ic_pct)
}
\arguments{
\item{fit}{A \code{drc} object}

\item{ic_pct}{numeric. The \%IC desired, where 25 would represent the
concentration at which growth was reduced by 25\% vs baseline}
}
\value{
A \code{tibble} with three columns - \code{ic_value}, \code{ic_std_err}, and
\code{ic_pct}
}
\description{
Get the IC value of a fit at a given percent
}
