% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pcr}
\alias{get_pcr}
\title{Get a variety of raw data}
\usage{
get_pcr(name, user_name = NULL, ...)
}
\arguments{
\item{name}{Character, the name of the file}

\item{user_name}{Optional character. Should be the name used for your
personal data directories in the SharePoint. If NULL, will use the
GBCI_USERNAME environmental variable.}

\item{...}{Additional arguments passed to \code{get_gbci}}
}
\description{
A collection of thin wrappers around \code{get_gbci}
}
