% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{get_plate}
\alias{get_plate}
\title{Create a plate with given well format}
\usage{
get_plate(format, no_border)
}
\arguments{
\item{format}{integer. How many wells should this plate have?}

\item{no_border}{logical. Will the user omit filling the edge wells of the
plate (to avoid edge effects)?}
}
\value{
A tibble with \code{col} (number, factor, denotes plate column from
left to right) and \code{row} (character, factor, denotes plate row from top to
bottom). If \code{no_border = TRUE}, the edges of the plate will not be included.
}
\description{
Create a plate with given well format
}
\keyword{internal}
