% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Get or make sample names}
\usage{
get_sample_names(data, has_names)
}
\arguments{
\item{data}{A data.frame}

\item{has_names}{logical. Should we expect sample names to be in the first
column?}
}
\value{
A vector of sample names
}
\description{
If the user did not supply sample names (\code{has_names = FALSE}), sample names
will be generated in the form of "Sample_n".
}
\keyword{internal}
