% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_pcr}
\alias{list_pcr}
\alias{list_nanodrop}
\alias{list_chemidoc}
\alias{list_licor}
\alias{list_evos}
\alias{list_qubit}
\alias{list_spectramax}
\title{List from a raw data folder}
\usage{
list_pcr(user_name = NULL, ...)

list_nanodrop(user_name = NULL, ...)

list_chemidoc(user_name = NULL, ...)

list_licor(user_name = NULL, ...)

list_evos(user_name = NULL, ...)

list_qubit(user_name = NULL, ...)

list_spectramax(user_name = NULL, ...)
}
\arguments{
\item{user_name}{Optional character. Should be the name used for your
personal data directories in the SharePoint. If NULL, will use the
GBCI_USERNAME environmental variable.}

\item{...}{Additional arguments passed to \code{list_gbci_dir}}
}
\description{
A collection of thin wrappers around \code{list_gbci_dir}.
}
