% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{make_curve}
\alias{make_curve}
\title{Make points for plotting based off a fit}
\usage{
make_curve(fit, drug_conc, length_out = 1000)
}
\arguments{
\item{fit}{a \code{drc} object}

\item{drug_conc}{numeric vector of drug concentrations. Assumes no 0s.}

\item{length_out}{number of points to generate. More points = smoother curve.}
}
\value{
A \code{data.frame} of x and y coordinates for a given fit
}
\description{
Make points for plotting based off a fit
}
