% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_protein.R
\name{make_qp_plate_view}
\alias{make_qp_plate_view}
\title{View the absorbances of an analyzed \code{qp} as they were on the plate}
\usage{
make_qp_plate_view(x)
}
\arguments{
\item{x}{The output of \code{qp()}}
}
\value{
a \code{ggplot}
}
\description{
View the absorbances of an analyzed \code{qp} as they were on the plate
}
