% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{make_qp_standard_plot}
\alias{make_qp_standard_plot}
\title{View an absorbance/concentration plot}
\usage{
make_qp_standard_plot(x)
}
\arguments{
\item{x}{The output of \code{qp()}}
}
\value{
a \code{ggplot}
}
\description{
View an absorbance/concentration plot
}
