% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{mtt}
\alias{mtt}
\alias{mtt.spectramax}
\title{Make a plot from MTT data}
\usage{
mtt(x, ...)

\method{mtt}{spectramax}(spectramax, condition_names, drug_conc)
}
\arguments{
\item{x}{A \code{data.frame}, \code{gp}, or \code{spectramax} object. See Details.}

\item{...}{Arguments passed to their repsective methods}
}
\value{
a \code{ggplot}
}
\description{
Make a plot from MTT data
}
\details{
If supplied with a \code{data.frame}, \code{mtt} will expect...

Using a \code{gp} object is a good idea if you have a 'non-standard' plate layout
(standard being each quarter of the plate is a condition)
}
