% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{mtt_calc}
\alias{mtt_calc}
\alias{mtt_calc.data.frame}
\alias{mtt_calc.spectramax}
\alias{mtt_calc.gplate}
\title{Calculate and fit MTT model from absorbance data}
\usage{
mtt_calc(x, ...)

\method{mtt_calc}{data.frame}(x, ic_pct = 50, ...)

\method{mtt_calc}{spectramax}(x, condition_names, drug_conc, ic_pct = 50, ...)

\method{mtt_calc}{gplate}(x, ic_pct = 50, ...)
}
\arguments{
\item{x}{A \code{data.frame}, \code{spectramax}, or \code{gp} object. See Details.}

\item{...}{Arguments passed to their respective methods}

\item{ic_pct}{numeric. The \%IC desired, where 25 would represent the
concentration at which growth was reduced by 25\% vs baseline}

\item{condition_names}{What to name each 'sector' of the plate}

\item{drug_conc}{a numeric vector containing drug concentrations of the
conditions, from left to right}
}
\value{
a \code{tibble}
}
\description{
Calculate and fit MTT model from absorbance data
}
\details{
If supplied with a \code{data.frame}, \code{mtt} will expect columns \code{condition},
\code{drug}, \code{nm562}, and \code{nm660}.

If supplied with a \code{spectramax} object, everything should be in place.

Using a \code{gp} object (NOT CURRENTLY SUPPORTED) is a good idea if you have a
'non-standard' plate layout (standard being each quarter of the plate is a
condition)
}
