% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{mtt_model}
\alias{mtt_model}
\title{Try to fit a logistic curve to MTT data}
\usage{
mtt_model(data)
}
\arguments{
\item{data}{a \code{data.frame} containing a \code{div} (dep var.) and \code{drug} (indep.
var) column.}
}
\value{
A \code{drc} object
}
\description{
Function will try to fit a 4 parameter log-logistic function. Constraints: Y
is bounded between 0 and 1.
}
\details{
If the model fails to fit the first time, it will try again without
constraints
}
