% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{mtt_plot}
\alias{mtt_plot}
\title{Plot MTT results}
\usage{
mtt_plot(mtt, plot_ics = FALSE)
}
\arguments{
\item{mtt}{a \code{data.frame} output from  \code{mtt_calc()}}
}
\value{
a \code{ggplot}
}
\description{
Plot MTT results
}
