% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{normalize_to_lowest_conc}
\alias{normalize_to_lowest_conc}
\title{Divide all differences by lowest concentration}
\usage{
normalize_to_lowest_conc(df)
}
\arguments{
\item{df}{a \code{data.frame} containing at \code{diff} (A562-A660), \code{drug} (numeric
conc of drug), and \code{mean} (mean of \code{diff} per condition and conc)}
}
\value{
a \code{tibble}
}
\description{
Divide all differences by lowest concentration
}
