% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_zero.R
\name{pad_zero}
\alias{pad_zero}
\title{Add leading 0 to sample names}
\usage{
pad_zero(sample_names)
}
\arguments{
\item{sample_names}{a vector of sample names}
}
\value{
A sample name with up to one zero padded
}
\description{
Add leading 0 to sample names
}
\keyword{internal}
