% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_path.R
\name{parse_path}
\alias{parse_path}
\title{Parse a SharePoint Path}
\usage{
parse_path(path)
}
\arguments{
\item{path}{SharePoint path copied and pasted from the icon next to the 'path' header from the file details menu in SharePoint}
}
\value{
a named list, including \code{host}, \code{site}, and \code{drive}
}
\description{
An internal function used to parse SharePoint paths into its constituitive parts
}
\examples{
path <- "https://mycompany.sharepoint.com/sites/mysite/mydrive/path/to/my/file.xls"
parse_path(path)
}
