% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_control}
\alias{pcr_control}
\title{(Re)calculate Delta Ct Mean}
\usage{
pcr_control(data, control_probe)
}
\arguments{
\item{data}{A dataset output from pcr_tidy}

\item{control_probe}{A probe to be used as an endogenous control (eg GAPDH)}
}
\value{
A tibble
}
\description{
(Re)calculate Delta Ct Mean
}
\examples{
dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "bladdr")

pcr_tidy(dat_path) |>
pcr_control("GAPDH")
}
