% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_libcalc.R
\name{pcr_libcalc}
\alias{pcr_libcalc}
\title{Calculate library pcr concentrations}
\usage{
pcr_libcalc(tidy_pcr, dil_factor = 1000)
}
\arguments{
\item{tidy_pcr}{a dataset run that has been run through \code{pcr_tidy()}}

\item{dil_factor}{the factor to which the libraries were diluted for pcr}
}
\value{
a list, containing the original dataframe, data on the standards, and data on the samples.
}
\description{
Calculate library pcr concentrations
}
\examples{

dat_path <- system.file("extdata", "untidy-standard-curve.xlsx", package = "bladdr")

pcr_tidy(dat_path) \%>\%
pcr_libcalc()

}
