% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_plan}
\alias{pcr_plan}
\title{Plan PCR experiment}
\usage{
pcr_plan(
  data,
  n_primers,
  format = 384,
  exclude_border = TRUE,
  primer_names = NULL,
  headless = TRUE,
  has_names = TRUE,
  make_report = FALSE,
  file_path = NULL
)
}
\arguments{
\item{data}{a data.frame, with samples as the first column (if \code{has_names = TRUE}) and RNA concentrations as the second (or first, if \code{has_names = FALSE})}

\item{n_primers}{integer. Number of primers to be used in the experiment.}

\item{format}{integer. 96 or 384 - the number of wells of the plate planned to be used}

\item{exclude_border}{logical. Should the border be excluded to avoid edge effects? Default is TRUE.}

\item{primer_names}{character vector. Names of primers.}

\item{headless}{logical. If FALSE, return invisible and redirect to shiny application.}

\item{has_names}{logical. Is the first column the names of the samples?}

\item{make_report}{logical. Should an HTML report be written?}

\item{file_path}{Optional path to where the report should be written, as well as the file name. Defaults to temp file.}
}
\value{
a list
}
\description{
Plan PCR experiment
}
