% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_plate_view}
\alias{pcr_plate_view}
\title{View sample plating layout}
\usage{
pcr_plate_view(tidy_pcr, fill = .data$target_name)
}
\arguments{
\item{tidy_pcr}{an output from the \code{pcr_tidy} function}

\item{fill}{the variable to use to fill the geom_tiles}
}
\value{
a ggplot
}
\description{
View sample plating layout
}
\examples{

dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "bladdr") |>
pcr_tidy() |>
pcr_plate_view()
}
