% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_plot}
\alias{pcr_plot}
\title{Plot qPCR results}
\usage{
pcr_plot(tidy_pcr)
}
\arguments{
\item{tidy_pcr}{an output from the \code{pcr_tidy} function, or some derivative thereof}
}
\value{
a ggplot
}
\description{
Plot qPCR results
}
\examples{

dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "bladdr")
pcr_tidy(dat_path) |>
pcr_rq("RD1") |>
pcr_plot()

}
