% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_tidy.R
\name{pcr_tidy}
\alias{pcr_tidy}
\title{Tidy a PCR Excel File}
\usage{
pcr_tidy(file_path = NULL)
}
\arguments{
\item{file_path}{Path to an excel file containing the results of a ddCT qPCR
run. If left blank, will open up interactive file chooser.}
}
\value{
Tidy results dataframe
}
\description{
Takes in a fresh results output from ddCT qPCR and converts it into a tidy
format. Useful for downstream analyses.
}
\examples{
dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "bladdr")

# Before tidying
dat_dirty <- readxl::read_excel(dat_path, sheet = "Results")
dat_dirty[1:10]

# After tidying
dat_clean <- pcr_tidy(dat_path)
dat_clean[1:10]
}
