% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{plate_template}
\alias{plate_template}
\title{Create a plate of given dimensions}
\usage{
plate_template(n_row, n_col)
}
\arguments{
\item{n_row}{integer. Number of rows plate should have.}

\item{n_col}{integer. Number of columns plate should have.}
}
\value{
A tibble with \code{col} (a factor of numbers with the lowest being the
first and the highest being the last) and \code{row} (a factor of characters in
reverse alphabetical order, to mimic ordering in real world plates when
plotting)
}
\description{
Create a plate of given dimensions
}
\keyword{internal}
