% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{qp_calc_dil}
\alias{qp_calc_dil}
\title{Calculate dilutions for an analyzed \code{qp} \code{list}}
\usage{
qp_calc_dil(x, target_conc, target_vol)
}
\arguments{
\item{x}{The output of \code{qp()}}

\item{target_conc}{Target concentration in (mg/mL) protein}

\item{target_vol}{Target volume in uL}
}
\value{
A list, where the \code{qp} item has volumes of lysate and volumes of H2O
to add.
}
\description{
Calculate dilutions for an analyzed \code{qp} \code{list}
}
