% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{qp_read}
\alias{qp_read}
\title{Read in data into a common quantify protein format}
\usage{
qp_read(x, ...)
}
\arguments{
\item{x}{A \code{gp}, \code{data.frame}/\code{tibble}, or character
path to a raw SPECTRAmax .xls(x)/.txt}

\item{...}{Unused}
}
\value{
A \code{gp}
}
\description{
Read in data into a common quantify protein format
}
