% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{rm_unassigned_wells}
\alias{rm_unassigned_wells}
\title{Remove unassigned wells}
\usage{
rm_unassigned_wells(df)
}
\arguments{
\item{df}{a \code{data.frame}}
}
\value{
a \code{tibble} without any conditions that are \code{NA}
}
\description{
Remove unassigned wells
}
