% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{sanitize_drug_conc}
\alias{sanitize_drug_conc}
\title{Convert 0 to a small enough equivalent}
\usage{
sanitize_drug_conc(drug_conc)
}
\arguments{
\item{drug_conc}{numeric vector of drug concentrations, in the order
which they appear in the section (left to right)}
}
\value{
a numeric vector of the same length and order as input
}
\description{
Convert 0 to a small enough equivalent
}
\details{
0 doesn't behave well with the fitting algorithm. This takes the
user-supplied drug concentrations and converts the 0 value (if it exists) to
a value low enough to approximate 0 without scaring the fitting algorithm too
much. Specifically, it converts 0 to:

\eqn{\frac{SecondSmallest}{ThirdSmallest^4}}
}
