% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_get.R
\name{sharepoint_get}
\alias{sharepoint_get}
\title{Download file from SharePoint}
\usage{
sharepoint_get(path, token)
}
\arguments{
\item{path}{a character vector. Obtained by going online to SharePoint, right clicking on file, clicking 'details', clicking 'more details' in the resultant side pane, then clicking the 'copy' button right next to the "Path" header.}

\item{token}{an AzureGraph token - see example.}
}
\value{
a tempfile path pointing to the location of the file downloaded
}
\description{
Download file from SharePoint
}
\examples{

# Getting token:
\dontrun{

gr <- AzureGraph::create_graph_login() # Will trigger a login
me <- gr$get_user("me")
token <- me$token
}

}
