% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summon_cell_data.R
\name{summon_cell_data}
\alias{summon_cell_data}
\title{Summon cell line data}
\usage{
summon_cell_data()
}
\value{
a DDS containing cell-line RNA expression
}
\description{
Summon cell line data
}
\details{
This function requires access to the GBCI SharePoint.
}
