% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{useable_plate_dims}
\alias{useable_plate_dims}
\title{Get dimensions of plate that are allowed to be filled}
\usage{
useable_plate_dims(plate)
}
\arguments{
\item{plate}{}
}
\value{
a named list, with integer values referring to the number of rows and
columns, respectively.
}
\description{
The 'useable' dimensions of the plate depends if the user has set
\code{exclude_border} to be \code{TRUE} or \code{FALSE}. This calculates plate dimensions
dependent on that.
}
\keyword{internal}
