% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_volcano.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Make a volcano plot from a dds results}
\usage{
volcano_plot(results, pval = 0.05, lfc = 0)
}
\arguments{
\item{results}{a \code{DESeqResults} object}

\item{pval}{the pvalue cutoff to consider significant}

\item{lfc}{the log fold change cutoff to consider significant}
}
\value{
a volcano plot
}
\description{
Make a volcano plot from a dds results
}
\examples{

DESeq2::makeExampleDESeqDataSet(n = 50, m = 3, betaSD = 1) |>
DESeq2::DESeq() |>
DESeq2::results() |>
volcano_plot(lfc = 1)

}
