% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gbci_file}
\alias{get_gbci_file}
\title{Get a file from the GBCI SharePoint}
\usage{
get_gbci_file(path)
}
\arguments{
\item{path}{Path to file on SharePoint}
}
\value{
Character. The local path of the downloaded file.
}
\description{
Get a file from the GBCI SharePoint
}
\details{
This function requires access to the SharePoint in the first place.
}
