% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtt}
\alias{mtt}
\title{Example data from an MTT}
\format{
A \code{spectramax} object
\describe{
\item{data}{a \code{gp} object containing absorbance values}
}
}
\usage{
mtt
}
\description{
A dataset including absorbance data at 562nm and 660nm
}
\details{
There are four conditions on this plate. The plate is divided into quadrants.
Concentrations of drug increase from left to right (0, 1nM, 10nM, 100nM, 1uM,
10uM).
}
\keyword{datasets}
