% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{mtt_calc}
\alias{mtt_calc}
\alias{mtt_calc.data.frame}
\alias{mtt_calc.spectramax}
\title{Make a plot from MTT data}
\usage{
mtt_calc(x, ...)

\method{mtt_calc}{data.frame}(x, ...)

\method{mtt_calc}{spectramax}(x, condition_names, drug_conc, ...)
}
\arguments{
\item{x}{A \code{data.frame}, \code{gp}, or \code{spectramax} object. See Details.}

\item{...}{Arguments passed to their repsective methods}
}
\value{
a \code{ggplot}
}
\description{
Make a plot from MTT data
}
\details{
If supplied with a \code{data.frame}, \code{mtt} will expect...

Using a \code{gp} object is a good idea if you have a 'non-standard' plate layout
(standard being each quarter of the plate is a condition)
}
