% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib_qc.R
\name{pcr_lib_qc}
\alias{pcr_lib_qc}
\title{Generate visual library prep pcr quality control}
\usage{
pcr_lib_qc(lib_calc_pcr)
}
\arguments{
\item{lib_calc_pcr}{an output from \code{pcr_lib_calc}}
}
\value{
a ggplot
}
\description{
Generate visual library prep pcr quality control
}
\examples{

dat_path <- system.file("extdata", "untidy-standard-curve.xlsx", package = "bladdr")

pcr_tidy(dat_path) \%>\%
        pcr_lib_calc() \%>\%
        pcr_lib_qc()
}
