% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{qp}
\alias{qp}
\title{Quantify protein concentration from a MicroBCA assay}
\usage{
qp(
  x,
  replicate_orientation = c("h", "v"),
  sample_names = NULL,
  remove_empty = TRUE,
  remove_outliers = c("all", "samples", "standards", "none")
)
}
\arguments{
\item{x}{A \code{spectramax}, \code{gp}, or \code{data.frame} object, or path to SPECTRAmax
.xls(x)/.txt file.}

\item{replicate_orientation}{Either 'h' or 'v' - see Details.}

\item{sample_names}{Character vector of sample names.}

\item{remove_empty}{Should wells that have less absorbance than the lowest
standard be dropped?}
}
\value{
a \code{tibble}
}
\description{
Quantify protein concentration from a MicroBCA assay
}
\details{
The standards must be in ascending concentration starting in the
upper left corner. Whether this is from from left to right or top to bottom
can be specified in 'replicate orientation'. Note that 'replicate
orientation' specified the direction that REPLICATES lie, NOT the direction
the samples flow (which will be opposite).
}
