% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt.R
\name{subtract_bg_and_get_mean}
\alias{subtract_bg_and_get_mean}
\title{Subtract background and calculate mean per drug and condition}
\usage{
subtract_bg_and_get_mean(df)
}
\arguments{
\item{df}{a \code{data.frame} containing columns \code{condition}, \code{drug}, \code{nm562}, and
\code{nm660}.}
}
\value{
a \code{tibble}
}
\description{
Subtract background and calculate mean per drug and condition
}
