% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_tufte}
\alias{theme_tufte}
\title{ggplot theme in the style of Edward Tufte}
\usage{
theme_tufte(font_size = 30)
}
\arguments{
\item{font_size}{numeric. Size of \code{element_text} font.}
}
\value{
A ggplot theme
}
\description{
ggplot theme in the style of Edward Tufte
}
